implementation of toSvg moved to object MultivarRangeQueryImpl
